requirements.txt
flask==2.2.2
Werkzeug==2.2.2
matplotlib==3.7.5
lifelines==0.27.8
matplotlib==3.7.5
numpy==1.24.4
Pillow==10.4.0
pytest==8.3.3
scikit_learn==1.3.2
seaborn==0.13.2
tqdm==4.64.0